+++ server/src/db/entities/index.ts
@@ -1,0 +1,11 @@
+export * from './base.entity';
+export * from './chat.entity';
+export * from './chattingchannel.entity';
+export * from './file.entity';
+export * from './groupmember.entity';
+export * from './meetingchannel.entity';
+export * from './reaction.entity';
+export * from './session.entity';
+export * from './thread.entity';
+export * from './user.entity';
+export * from './workgroup.entity';
