+++ server/src/db/repositories/chat.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { Chat } from '../entities/chat.entity';
+import { Chat } from '../entities';
 
 @EntityRepository(Chat)
 export class ChatRepository extends Repository<Chat> {
