+++ server/src/db/repositories/thread.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { Thread } from '../entities/thread.entity';
+import { Thread } from '../entities';
 
 @EntityRepository(Thread)
 export class ThreadRepository extends Repository<Thread> {
