+++ server/src/db/repositories/user.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { User } from '../entities/user.entity';
+import { User } from '../entities';
 
 @EntityRepository(User)
 export class UserRepository extends Repository<User> {
